Refactor React components into separate files
